Factor out pdf parsing code
