+++ dle/data/management/commands/load_ema_data.py
@@ -12,7 +12,6 @@ from django.core.management.base import BaseCommand, CommandError
 from django.db import IntegrityError
 
 import pandas as pd
-import pdfplumber
 import requests
 from bs4 import BeautifulSoup
 from Levenshtein import distance as levdistance
@@ -20,6 +19,8 @@ from requests.exceptions import ChunkedEncodingError
 
 from data.models import DrugLabel, LabelProduct, ProductSection
 from users.models import MyLabel
+
+from .pdf_parsing_helper import get_pdf_sections, read_pdf
 
 
 logger = logging.getLogger(__name__)
@@ -331,86 +332,8 @@ class Command(BaseCommand):
         else:
             return self.centers[ix]
 
-    # function: input text, output list of section headers and content
-    def get_smpc_sections(self, text):
-        idx, headers, sections = [], [], []
-        for i, line in enumerate(text):
-            if re.match(r"^[0-9]+\.[0-9]*\s+.*[A-Z].*", line):
-                idx += [i]
-                headers += [line.strip()]
-
-        # in headers, must increment or restart, and not end in punctuation
-        idx_valid, headers_valid = [idx[0]], [headers[0]]
-        for n in range(1, len(headers)):
-            prev = float(headers[n - 1].split()[0])
-            curr = float(headers[n].split()[0])
-            lastchar = headers[n].strip()[-1].lower()
-            valid = (prev < curr <= prev + 1) or (curr == 1)
-            valid = valid and (lastchar in "qwertyuiopasdfghjklzxcvbnm()")
-            if valid:
-                idx_valid.append(idx[n])
-                headers_valid.append(headers[n])
-        idx, headers = idx_valid, headers_valid
-
-        for n, h in enumerate(headers):
-            if (n + 1) < len(headers):
-                contents = text[idx[n] + 1 : idx[n + 1]]
-            else:
-                contents = text[idx[n] + 1 :]
-            sections += ["\n".join(contents)]
-
-        return headers, sections
-
-    # helper function for pdfplumber
-    def remove_tables(self, page):
-        ts = {"vertical_strategy": "lines", "horizontal_strategy": "lines"}
-        bboxes = [table.bbox for table in page.find_tables(table_settings=ts)]
-
-        def not_within_bboxes(obj):
-            # Check if the object is in any of the table's bbox.
-            def obj_in_bbox(_bbox):
-                # See https://github.com/jsvine/pdfplumber/blob/stable/pdfplumber/table.py#L404
-                v_mid = (obj["top"] + obj["bottom"]) / 2
-                h_mid = (obj["x0"] + obj["x1"]) / 2
-                x0, top, x1, bottom = _bbox
-                return (h_mid >= x0) and (h_mid < x1) and (v_mid >= top) and (v_mid < bottom)
-
-            return not any(obj_in_bbox(__bbox) for __bbox in bboxes)
-
-        return page.filter(not_within_bboxes)
-
-    # helper function for pdfplumber
-    def remove_margins(self, page, dpi=72, size=0.7):
-        # strip 0.7 inches from top and bottom (page numbers, header text if any), A4 is 8.25 x 11.75
-        # syntax is page.crop((x0, top, x1, bottom))
-        w = float(page.width) / dpi
-        h = float(page.height) / dpi
-        return page.crop((0, (size) * dpi, w * dpi, (h - size) * dpi))
-
-    # function: input file, output text of annex 1
-    def read_smpc(self, filename, no_blanks=False, no_tables=False):
-        text = []
-        with pdfplumber.open(filename) as pdf:
-            for page in pdf.pages:
-                page = self.remove_margins(page)
-
-                if no_tables:
-                    page = self.remove_tables(page)
-
-                page_text = page.extract_text().split("\n")
-                text += page_text
-
-        annex_lines = [re.match(r".*ANNEX\s+I.*", line) is not None for line in text]
-        annex_index = [i for i, v in enumerate(annex_lines) if v]
-
-        text = text[annex_index[0] : annex_index[1]]
-        if no_blanks:
-            text = [line for line in text if not line.isspace()]
-
-        return text
-
     def process_ema_file(self, ema_file, lp, pdf_url=""):
-        text = self.read_smpc(ema_file)
+        text = read_pdf(ema_file)
 
         try:
             info = {}
@@ -419,7 +342,7 @@ class Command(BaseCommand):
             info["metadata"] = row.iloc[0].apply(str).to_dict()
 
             label_text = {}  # next level = product page w/ metadata
-            headers, sections = self.get_smpc_sections(text)
+            headers, sections = get_pdf_sections(text, pattern=r"^[0-9]+\.[0-9]*\s+.*[A-Z].*")
 
             for h, s in zip(headers, sections):
                 header = self.get_fixed_header(h)
